% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellspell_config.R
\name{get_default_dict_hunspell}
\alias{get_default_dict_hunspell}
\title{Get default hunspell dictionary.}
\usage{
get_default_dict_hunspell(hunspell_dicts = NULL)
}
\arguments{
\item{hunspell_dicts}{(character or \code{NULL}) A list of hunspell dictionaries
as a character vector.
If \code{NULL}, the function gets a list of installed dictionaries.}
}
\value{
String with default language or \code{""}.
}
\description{
Get short name of a Hunspell dictionary for spellchecking.
Note that this name usually contains underscores and not hyphens,
e.g., \code{"en_GB"}.
}
\details{
Function tries to match the default RStudio spelling language first.
If it is not on the list, then English/Great Britain (\code{"en_GB"}) is the second choice.
If it is not present, the first language in the list is selected.
If the list is empty, empty string \code{""} is returned.
}
\keyword{internal}
