% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellspell_config.R
\name{get_default_dict_languagetool}
\alias{get_default_dict_languagetool}
\title{Get the default languagetool dictionary.}
\usage{
get_default_dict_languagetool(languagetool_dicts = NULL)
}
\arguments{
\item{languagetool_dicts}{(character or \code{NULL}) A list of languagetool dictionaries
as a character vector.
If \code{NULL}, the function gets a list of installed dictionaries.}
}
\value{
String with default language or \code{""}.
}
\description{
Get short name of a \code{languagetool} dictionary for grammar checking.
Note that this name usually contains hyphens and not underscores
(e.g., \code{"en-GB"}).
}
\details{
Function tries to match the default RStudio spelling language first.
If no exact match is found, tries to find the match of the first two letters.
The next choice is English/Great Britain (\code{"en-GB"}) if it exists in the lists.
Then first language in the list.
If the list is empty, empty string \code{""} is returned.
}
