% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellspell.R, R/wellspell_config.R
\docType{package}
\name{wellspell.addin-package}
\alias{wellspell.addin}
\alias{wellspell.addin-package}
\alias{spellcheck}
\alias{get_config}
\alias{is_config}
\alias{rm_config}
\alias{set_config}
\title{wellspell.addin: Quick Spellcheck Addin for Rstudio}
\usage{
spellcheck()

get_config()

is_config()

rm_config()

set_config()
}
\description{
test
}
\author{
\strong{Maintainer}: Clemens Schmid \email{clemens@nevrome.de}

}
