% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellspell.R, R/wellspell_config.R
\docType{package}
\name{wellspell.addin-package}
\alias{wellspell.addin}
\alias{wellspell.addin-package}
\alias{spellcheck}
\alias{gramcheck}
\alias{get_config}
\alias{is_config}
\alias{rm_config}
\alias{set_config}
\title{wellspell.addin}
\usage{
spellcheck()

gramcheck()

get_config()

is_config()

rm_config()

set_config()
}
\description{
wellspell is an RStudio Addin to quickly highlight words with 
spelling errors in text documents. It employs the hunspell spell checking engine
via the \href{https://github.com/ropensci/hunspell}{hunspell package} to do this
to do this.

To use it, you can select an arbitrary amount of text in an text document 
in RStudio (e.g. a markdown, latex or html document) and run \code{spellcheck()}.
As the function is registered as an RStudio Addin, it's possible to run it from
the Addins dialogue or even with a keyboard shortcut (e.g. Ctrl+Alt+7). 

At the first run in a new environment, \code{spellcheck()} will call 
\code{set_config()}, which is another Addin with a minimalistic user interface. 
It allows you to set two environment variables \code{wellspell_language} and
\code{wellspell_format}. These are used to configure \code{hunspell::hunspell()}.
\code{wellspell_language} is fed to \code{hunspell(dict = dictionary(lang = ...))}
and \code{wellspell_format} to \code{hunspell(format = ...)}. 

If the environment variables are set, \code{spellcheck()} selects and thereby 
highlights all words identified as wrong by hunspell.

The additional functions \code{get_config()}, \code{is_config()} and \code{rm_config}
are for dealing with the environment variables and usually don't have to be called 
directly. 

\strong{How to install dictionaries for other languages}
  
RStudio's default installation includes English dictionaries for the US, UK, Canada, 
and Australia. In addition, dictionaries for many other languages can be installed. 
To add these dictionaries, go to the \emph{Spelling} pane of the \emph{Options} dialog, and 
select \emph{Install More Languages...} from the language dictionary select box. This will 
download and install all of the available languages. Further instructions can be found 
\href{https://support.rstudio.com/hc/en-us/articles/200551916-Spelling-Dictionaries}{here}. 
If this doesn't work or the relevant languages are not in the default selection you can 
install languages by copying the dictionary files (.dic + .aff) to one of these locations: 
\code{hunspell::dicpath()}.
}
\author{
\strong{Maintainer}: Clemens Schmid \email{clemens@nevrome.de}

}
